rename folder to tabs
