+++ src/routes/(app)/collection/[...collectionName]/tabs/CollectionForm.svelte
