+++ src/routes/(app)/collection/[...collectionName]/tabs/CollectionPermission.svelte
