+++ src/routes/(app)/collection/[...collectionName]/tabs/CollectionWidget.svelte
