+++ src/routes/(app)/collection/[...collectionName]/tabs/TopTabs.svelte
